backend/requirements/dev.txt
# Core FastAPI & Server
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# Database (Async)
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.25.0 # Driver async pour PostgreSQL avec SQLAlchemy

# Pydantic & Validation
pydantic>=2.0.0 
email-validator>=1.1 # Requis pour Pydantic EmailStr
sqlmodel>=0.0.14 # Ajout de SQLModel

# Authentication & Security
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0

# Templating
Jinja2>=3.0

# PDF Generation
reportlab>=3.6

# LLM (LangChain)
langchain-community>=0.0.20 # Version minimale, ajuster si besoin
langchain-core>=0.1.20 # Version minimale, ajuster si besoin

# Configuration
python-dotenv>=1.0.0 # Pour charger les variables de .env

# CRUD Helper (Optionnel, était présent)
fastcrud>=0.10.0

# Form Data / File Uploads
python-multipart>=0.0.5

pytest
httpx

pytest-asyncio

aiosqlite
# Ajouter d'autres dépendances spécifiques si nécessaire (ex: pydantic, email_validator, etc.)
# Note: Pydantic est souvent inclus avec FastAPI, mais on peut le spécifier si besoin d'une version précise

# Note: Pydantic est souvent inclus avec FastAPI, mais on peut le spécifier si besoin d'une version précise 